move files to their dirs
