+++ Sources/AdaEngine/Scene/2D/Sprite/SpriteRenderSystem.swift
@@ -5,16 +5,22 @@
 //  Created by v.prusakov on 5/10/22.
 //
 
-struct SpriteRenderSystem: System {
+public struct SpriteRenderSystem: System {
     
-    static var dependencies: [SystemDependency] = [.before(Physics2DSystem.self)]
+    public static var dependencies: [SystemDependency] = [.before(Physics2DSystem.self)]
     
     static let cameras = EntityQuery(where:
             .has(Camera.self) &&
             .has(VisibleEntities.self) &&
-            .has(Transform.self) &&
             .has(RenderItems<Transparent2DRenderItem>.self)
     )
+    
+    struct SpriteVertexData {
+        let position: Vector4
+        let color: Color
+        let textureCoordinate: Vector2
+        let textureIndex: Int
+    }
     
     static let quadPosition: [Vector4] = [
         [-0.5, -0.5,  0.0, 1.0],
@@ -26,7 +32,7 @@ struct SpriteRenderSystem: System {
     let quadRenderPipeline: RenderPipeline
     let gpuWhiteTexture: Texture2D
     
-    init(scene: Scene) {
+    public init(scene: Scene) {
         let device = RenderEngine.shared
         
         var samplerDesc = SamplerDescriptor()
@@ -64,15 +70,11 @@ struct SpriteRenderSystem: System {
         self.gpuWhiteTexture = Texture2D(image: image)
     }
     
-    func update(context: UpdateContext) {
+    public func update(context: UpdateContext) {
         context.scene.performQuery(Self.cameras).forEach { entity in
-            var (camera, _, visibleEntities, renderItems) = entity.components[Camera.self, Transform.self, VisibleEntities.self, RenderItems<Transparent2DRenderItem>.self]
+            var (camera, visibleEntities, renderItems) = entity.components[Camera.self, VisibleEntities.self, RenderItems<Transparent2DRenderItem>.self]
             
             if !camera.isActive {
-                return
-            }
-            
-            if case .window(let id) = camera.renderTarget, id == .empty {
                 return
             }
             
@@ -93,9 +95,9 @@ struct SpriteRenderSystem: System {
         let sprites = visibleEntities.filter {
             $0.components.has(SpriteComponent.self)
         }
-        .sorted { lhs, rhs in
-            lhs.components[Transform.self]!.position.z < rhs.components[Transform.self]!.position.z
-        }
+            .sorted { lhs, rhs in
+                lhs.components[Transform.self]!.position.z < rhs.components[Transform.self]!.position.z
+            }
         
         var spriteVerticies = [SpriteVertexData]()
         spriteVerticies.reserveCapacity(MemoryLayout<SpriteVertexData>.stride * sprites.count)
@@ -120,9 +122,7 @@ struct SpriteRenderSystem: System {
             }
             
             if let sprite = entity.components[SpriteComponent.self] {
-                
                 // Select a texture index for draw
-                
                 let textureIndex: Int
                 
                 if let texture = sprite.texture {
@@ -166,6 +166,7 @@ struct SpriteRenderSystem: System {
                 )
             }
             
+            // TODO: Should be in debug render system
             if scene.debugOptions.contains(.showBoundingBoxes) {
                 if let bounding = entity.components[BoundingComponent.self] {
                     guard case .aabb(let aabb) = bounding.bounds else {
@@ -219,17 +220,17 @@ struct SpriteRenderSystem: System {
         let indicies = Int(indeciesCount * 4)
         
         var quadIndices = [UInt32].init(repeating: 0, count: indicies)
-
+        
         var offset: UInt32 = 0
         for index in stride(from: 0, to: indicies, by: 6) {
             quadIndices[index + 0] = offset + 0
             quadIndices[index + 1] = offset + 1
             quadIndices[index + 2] = offset + 2
-
+            
             quadIndices[index + 3] = offset + 2
             quadIndices[index + 4] = offset + 3
             quadIndices[index + 5] = offset + 0
-
+            
             offset += 4
         }
         
@@ -251,13 +252,6 @@ struct SpriteRenderSystem: System {
     }
 }
 
-struct SpriteVertexData {
-    let position: Vector4
-    let color: Color
-    let textureCoordinate: Vector2
-    let textureIndex: Int
-}
-
 struct SpriteDataComponent: Component {
     let vertexBuffer: VertexBuffer
     let indexArray: RID
@@ -265,104 +259,4 @@ struct SpriteDataComponent: Component {
 
 public struct BatchComponent: Component {
     public var textures: [Texture2D]
-}
-
-struct BatchTransparent2DItemsSystem: System {
-
-    static let query = EntityQuery(where: .has(RenderItems<Transparent2DRenderItem>.self))
-
-    init(scene: Scene) { }
-
-    func update(context: UpdateContext) {
-        context.scene.performQuery(Self.query).forEach { entity in
-            guard let renderItems = entity.components[RenderItems<Transparent2DRenderItem>.self] else {
-                return
-            }
-            
-            let items = renderItems.items
-            var batchedItems: [Transparent2DRenderItem] = []
-            batchedItems.reserveCapacity(items.count)
-            
-            if var currentItem = items.first {
-                
-                for nextItemIndex in 1..<items.count {
-                    let nextItem = items[nextItemIndex]
-                    
-                    if tryToAddBatch(to: &currentItem, from: nextItem) == false {
-                        batchedItems.append(currentItem)
-                        currentItem = nextItem
-                    }
-                }
-                
-                batchedItems.append(currentItem)
-            }
-            
-            entity.components[RenderItems<Transparent2DRenderItem>.self] = RenderItems(items: batchedItems)
-        }
-    }
-    
-    private func tryToAddBatch(to currentItem: inout Transparent2DRenderItem, from otherItem: Transparent2DRenderItem) -> Bool {
-        guard let batch = currentItem.batchRange, let otherBatch = otherItem.batchRange else {
-            return false
-        }
-        
-        if otherItem.batchEntity !== currentItem.batchEntity {
-            return false
-        }
-        
-        if batch.upperBound == otherBatch.lowerBound {
-            currentItem.batchRange = batch.lowerBound ..< otherBatch.upperBound
-        } else if batch.lowerBound == otherBatch.upperBound {
-            currentItem.batchRange = otherBatch.lowerBound ..< batch.upperBound
-        } else {
-            return false
-        }
-        
-        return true
-    }
-}
-
-struct SpriteDrawPass: DrawPass {
-    
-    struct SpriteViewUniform {
-        let viewMatrix: Transform3D
-    }
-    
-    let uniformBufferSet: UniformBufferSet
-    
-    init() {
-        uniformBufferSet = RenderEngine.shared.makeUniformBufferSet()
-        uniformBufferSet.initBuffers(for: SpriteViewUniform.self, count: 3, binding: 1, set: 0)
-    }
-    
-    func render(in context: Context, item: Transparent2DRenderItem) throws {
-        guard let spriteData = context.entity.components[SpriteDataComponent.self] else {
-            return
-        }
-        
-        guard let batchSprite = item.batchEntity.components[BatchComponent.self] else {
-            return
-        }
-        
-        guard let cameraViewUniform = context.view.components[ViewUniform.self] else {
-            return
-        }
-        
-        guard let count = item.batchRange?.count, count > 0 else {
-            return
-        }
-        
-        let uniform = uniformBufferSet.getBuffer(binding: 1, set: 0, frameIndex: context.device.currentFrameIndex)
-        uniform.setData(SpriteViewUniform(viewMatrix: cameraViewUniform.viewProjectionMatrix))
-        
-        batchSprite.textures.enumerated().forEach { (index, texture) in
-            context.drawList.bindTexture(texture, at: index)
-        }
-        context.drawList.appendUniformBuffer(uniform)
-        context.drawList.appendVertexBuffer(spriteData.vertexBuffer)
-        context.drawList.bindIndexArray(spriteData.indexArray)
-        context.drawList.bindRenderPipeline(item.renderPipeline)
-        
-        context.drawList.drawIndexed(indexCount: count, instancesCount: 1)
-    }
 }
