+++ Sources/AdaEngine/Scene/ScenePlugins/RenderPlugins/Scene2DPlugin.swift
@@ -40,21 +40,6 @@ public struct Scene2DPlugin: ScenePlugin {
     }
 }
 
-public struct ClearTransparent2DRenderItemsSystem: System {
-    
-    public static var dependencies: [SystemDependency] = [.before(SpriteRenderSystem.self)]
-    
-    static let query = EntityQuery(where: .has(RenderItems<Transparent2DRenderItem>.self))
-    
-    public init(scene: Scene) { }
-    
-    public func update(context: UpdateContext) {
-//        context.scene.performQuery(Self.query).forEach { entity in
-//            entity.components[RenderItems<Transparent2DRenderItem>.self]?.items.removeAll()
-//        }
-    }
-}
-
 public struct Main2DRenderNode: RenderNode {
     
     public static let name: String = "main_pass_2d"
@@ -123,82 +108,6 @@ public struct Main2DRenderNode: RenderNode {
     }
 }
 
-public struct RenderItems<T: RenderItem>: Component {
-    public var items: [T]
-    
-    public init(items: [T] = []) {
-        self.items = items
-    }
-    
-    public mutating func sort() {
-        self.items.sort(by: { $0.sortKey < $1.sortKey })
-    }
-    
-    public func sorted() -> Self {
-        var value = self
-        value.items.sort(by: { $0.sortKey < $1.sortKey })
-        return value
-    }
-    
-    public func render(_ drawList: DrawList, world: World, view: Entity) throws {
-        for item in self.items {
-            guard let drawPass = DrawPassStorage.getDrawPass(for: item) else {
-                continue
-            }
-            
-            let context = RenderContext(
-                device: .shared,
-                entity: item.entity,
-                world: world,
-                view: view,
-                drawList: drawList
-            )
-            
-            try drawPass.render(in: context, item: item)
-        }
-    }
-    
-    private func render(in context: RenderContext, item: T, drawPass: any DrawPass<T>) throws {
-        try drawPass.render(in: context, item: item)
-    }
-}
-
-public struct DrawPassId: Equatable, Hashable {
-    let id: Int
-}
-
-public protocol RenderItem {
-    
-    associatedtype SortKey: Comparable
-    
-    var entity: Entity { get }
-    var drawPassId: DrawPassId { get }
-    var sortKey: SortKey { get }
-}
-
-public struct RenderContext {
-    public let device: RenderEngine
-    public let entity: Entity
-    public let world: World
-    public let view: Entity
-    public let drawList: DrawList
-}
-
-public protocol DrawPass<Item> {
-    
-    associatedtype Item: RenderItem
-    typealias Context = RenderContext
-    
-    func render(in context: Context, item: Item) throws
-}
-
-public extension DrawPass {
-    /// Return identifier of draw pass based on DrawPass.Type
-    @inline(__always) static var identifier: DrawPassId {
-        DrawPassId(id: Int(bitPattern: ObjectIdentifier(self)))
-    }
-}
-
 public struct Transparent2DRenderItem: RenderItem {
     public var entity: Entity
     public var batchEntity: Entity
@@ -206,57 +115,4 @@ public struct Transparent2DRenderItem: RenderItem {
     public var renderPipeline: RenderPipeline
     public var sortKey: Float
     public var batchRange: Range<Int32>?
-}
-
-public enum DrawPassStorage {
-    
-    private static var draws: [DrawPassId: any DrawPass] = [:]
-    
-    private static let lock: NSLock = NSLock()
-    
-    public static func getDrawPass<I: RenderItem>(for item: I) -> AnyDrawPass<I>? {
-        lock.lock()
-        defer { lock.unlock() }
-        guard let drawPass = draws[item.drawPassId] else {
-            return nil
-        }
-        
-        return AnyDrawPass(drawPass)
-    }
-    
-    public static func setDrawPass<T: DrawPass>(_ drawPass: T) {
-        lock.lock()
-        defer { lock.unlock() }
-        
-        let key = T.identifier
-        draws[key] = drawPass
-    }
-}
-
-public struct SpritePlugin: ScenePlugin {
-    
-    public init() {}
-    
-    public func setup(in scene: Scene) {
-        scene.addSystem(SpriteRenderSystem.self)
-        scene.addSystem(BatchTransparent2DItemsSystem.self)
-        
-        let spriteDraw = SpriteDrawPass()
-        DrawPassStorage.setDrawPass(spriteDraw)
-    }
-}
-
-public struct AnyDrawPass<T: RenderItem>: DrawPass {
-    
-    private var render: (Context, Any) throws -> Void
-    
-    public init<Value: DrawPass>(_ base: Value) {
-        self.render = { context, item in
-            try base.render(in: context, item: item as! Value.Item)
-        }
-    }
-    
-    public func render(in context: Context, item: T) throws {
-        try render(context, item)
-    }
 }
